Filter pages WIP
Filtering the pages isn't too hard, and works.

The problem I ran in to here is that there is currently no way to filter
the browser_stat, size_stat, and location_stat by path/title. Guess I'll
have to add those columns, but don't really fancy making this data more
complicated.

Alternative might be to just filter this data and display these stats
as-is; I think that might actually be okay. Need to think about it.

Issue #62